refactor: 프론트 홈 패키지 구조 개선
